search: normalize with simple 0..1 linear mapping
